move migrateTo4K.cpp into helpers/nrf52
